refactor: use wizard container header, replace toggle button with switch and fix finish button set buy now
