--- src/features/domain-settings/DomainSettingsForm/ui/FormHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './FormHeader';
