--- src/features/domain-settings/DomainSettingsForm/ui/index.ts
@@ -1,2 +1,0 @@
-export * from './FormHeader';
-export * from './Switch';
