+++ src/features/ui/Switch/Switch.module.scss
@@ -1,6 +1,6 @@
-@import '../../../../../../node_modules/@zero-tech/zui/styles/_animations.scss';
-@import '../../../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
-@import '../../../../../../node_modules/@zero-tech/zui/styles/_typography.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_animations.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_typography.scss';
 
 .Container {
 	display: flex;
