--- src/features/ui/ToggleButton/index.ts
@@ -1,1 +1,0 @@
-export * from './ToggleButton';
