+++ src/features/ui/index.ts
@@ -8,5 +8,5 @@ export * from './FormUserBalance';
 export * from './IconButton';
 export * from './OptionLabel';
 export * from './StatsList';
+export * from './Switch';
 export * from './TableCard';
-export * from './ToggleButton';
